+++ include/builtins.h
@@ -6,7 +6,7 @@
 /*   By: gpaeng <gpaeng@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/02 14:53:53 by gpaeng            #+#    #+#             */
-/*   Updated: 2021/11/11 17:37:01 by gpaeng           ###   ########.fr       */
+/*   Updated: 2021/11/13 13:38:27 by gpaeng           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,8 @@ char *ft_strdup(const char *str);
 char *ft_strjoin(char const *s1, char const *s2);
 size_t ft_strlcpy(char *dest, const char *src, size_t size);
 int	ft_strncmp(const char *s1, const char *s2, size_t n);
+char *ft_set_malloc(int size);
+void    ft_free_arr(char **arr);
 
 /*
 **ft_echo.c
