+++ src/utils/ft_cnt_arg.c
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_cnt_arg.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gpaeng <gpaeng@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/11/13 13:38:55 by gpaeng            #+#    #+#             */
+/*   Updated: 2021/11/13 13:39:08 by gpaeng           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../include/builtins.h"
+
+int     ft_cnt_arg(char *args[])
+{
+    int i;
+
+    i = 0;
+    while (args[i])
+    {
+        i++;
+    }
+    return (i);
+}
